# Conflicts:
#	app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/infra/gui/FinalizaCadastroActivity.java
#	app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/infra/gui/LoginActivity.java
#	app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/infra/gui/MapsActivity.java
#	app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/infra/negocio/ApiRequestService.java

+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/animal/dominio/Animal.java
@@ -55,7 +55,7 @@ public class Animal {
         return peso;
     }
 
-    public void setPeso(int peso) {
+    public void setPeso(float peso) {
         this.peso = peso;
     }
